fix(i18n): show both experience values on Linux
The "EXP:%I64d / %I64d" strings use MSVC's %I64d. glibc reads the I flag as
"locale digits" and 64 as a field width, so the character window and HUD
experience tooltip padded each value to 64 columns and the 128-character
buffer cut off the next-level experience.

Use %llu in every localization and pass unsigned long long explicitly at
both call sites. A test formats the string in every locale.

+++ src/source/UI/Character/CharacterInfoWindow.cpp
@@ -515,7 +515,8 @@ void mu::ui::window::CCharacterInfoWindow::BuildTableTexts()
     wchar_t strPoint[128];
 
     mu_swprintf(strLevel, I18N::Game::LevelUResetsU, CharacterAttribute->Level, CharacterAttribute->Resets);
-    mu_swprintf(strExp, I18N::Game::EXPI64dI64d, CharacterAttribute->Experience, CharacterAttribute->NextExperience);
+    mu_swprintf(strExp, I18N::Game::EXPI64dI64d, static_cast<unsigned long long>(CharacterAttribute->Experience),
+                static_cast<unsigned long long>(CharacterAttribute->NextExperience));
 
     if (CharacterAttribute->Level > 9)
     {
